% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{run_peak_merging}
\alias{run_peak_merging}
\title{Run the full mumerge peak merging workflow}
\usage{
run_peak_merging(files_path, pattern = "bed", ncores = 1, chunk_size = 500)
}
\arguments{
\item{files_path}{Path to the directory containing BED/XLS peak files.}

\item{ncores}{Number of cores to use for parallel processing.}

\item{chunk_size}{Number of regions to process in each parallel chunk.}
}
\value{
A tibble with the final merged peaks (seqnames, start, end).
}
\description{
This is the main wrapper function that takes file paths as input,
reads the data, processes it in parallel, and returns the final set
of merged peaks.
}
